refactor: turn require admin into a helper
n
