+++ src/features/terminal/components/terminal-history/index.tsx
