+++ src/features/terminal/components/terminal-history/terminal-badge-small.tsx
