+++ src/features/terminal/components/terminal-select.tsx
